+++ src/app/(app)/layout.js
@@ -2,7 +2,7 @@
 import "@/app/globals.css";
 import Header from "@/components/header";
 import Footer from "@/components/Footer";
-
+import SessionProviderWrapper from "@/components/SessionProviderWrapper";
 
 export const metadata = {
   title: "DevInsights",
@@ -13,15 +13,16 @@ export default function RootLayout({ children }) {
   return (
     <html
       lang="en"
-      className={` bg-gray-900`}
     >
       <head>
         <meta name="viewport" content="width=device-width, initial-scale=1" />
       </head>
-      <body className={` antialiased`}>
-        <Header />
-        {children}
-        <Footer />
+      <body className={` antialiased bg-fixed`}>
+        <SessionProviderWrapper>
+          <Header />
+          {children}
+          <Footer />
+        </SessionProviderWrapper>
       </body>
     </html>
   );