+++ src/app/admin/(admin)/layout.jsx
@@ -1,6 +1,8 @@
 import "@/app/globals.css";
 import AdminApp from '@/components/AdminApp.jsx'
 import { Toaster } from "react-hot-toast";
+import { auth } from "@/auth";
+import { redirect } from "next/navigation";
 
 
 
@@ -10,7 +12,10 @@ export const metadata = {
 };
 
 export default async function AuthLayout({ children }) {
-  
+  const session = await auth();
+  if (!session) {
+    redirect('/admin/login');
+  }
   return (
     <html lang="en" className={``}>
       <head>